requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
sqlalchemy==2.0.30
openai==1.30.0
yt-dlp==2024.5.27
python-dotenv==1.0.1
aiofiles==23.2.1
